Add graceful exit to email and process servers
- email and process servers now check for a kill file in /sgm
(/sgm/killE, or /sgm/kill1, /sgm/kill2, ...) after
handling each subjob, exiting if they find it

- the runMotusXXXServer.sh scripts don't restart a server if
it stops due to a kill file

- add a "-g" option to the killMotusXXXServer scripts to permit graceful
exit via killfiles

+++ inst/scripts/killMotusEmailServer.sh
@@ -1,9 +1,22 @@
 #!/bin/bash
 #
 # kill the motus email server if it is running
+# Specify "-g" ("graceful") to let it finish its current subjob first.
+
+KILLFILE=/sgm/killE
+GRACEFUL=""
+if [[ "$1" == "-g" ]]; then
+    GRACEFUL=1
+    shift
+fi
+
 PIDFILE=/sgm/emailServer.pid
 PID=`cat $PIDFILE`
 if [[ "$PID" != "" ]]; then
-    pkill -g $PID
+    if [[ $GRACEFUL ]]; then
+        touch /sgm/killE
+    else
+        pkill -g $PID
+    fi
     rm -f $PIDFILE
 fi
